+++ packages/orq-rc/src/models/operations/findonebykeypromptsnippet.ts
@@ -4379,7 +4379,7 @@ export const FindOneByKeyPromptSnippetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-02-11T11:00:12.130Z",
+    "2025-02-11T12:04:07.091Z",
   ).transform(v => new Date(v)),
   type: FindOneByKeyPromptSnippetType$inboundSchema,
   versions: z.array(
@@ -4430,7 +4430,7 @@ export const FindOneByKeyPromptSnippetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string(),
   updatedById: z.string(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-02-11T11:00:12.130Z"))
+  updated: z.date().default(() => new Date("2025-02-11T12:04:07.091Z"))
     .transform(v => v.toISOString()),
   type: FindOneByKeyPromptSnippetType$outboundSchema,
   versions: z.array(