+++ packages/orq-rc/RELEASES.md
@@ -399,3 +399,13 @@ Based on:
 - [typescript v3.1.0-rc.60] packages/orq-rc
 ### Releases
 - [NPM v3.1.0-rc.60] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.60 - packages/orq-rc
+
+## 2025-02-11 12:05:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.488.0 (2.506.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.1.0-rc.61] packages/orq-rc
+### Releases
+- [NPM v3.1.0-rc.61] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.61 - packages/orq-rc